+++ client/src/routes/Routing.routes.tsx
@@ -1,17 +1,17 @@
 import { SvgIconProps } from '@material-ui/core/SvgIcon';
 import {
   Account as StudentIcon,
-  BadgeAccount as UserIcon,
   AccountMultiple as TeamIcon,
   AccountMultipleCheck as AttendancesIcon,
+  BadgeAccount as UserIcon,
   Book as EnterPointsIcon,
   Comment as PresentationIcon,
   File as SheetIcon,
-  TextBox as ScheinexamPointsIcon,
-  TextBoxMultiple as ScheinexamManagementIcon,
   Login as LoginIcon,
   ScriptText as ScheincriteriaIcon,
   Teach as TutorialIcon,
+  TextBox as ScheinexamPointsIcon,
+  TextBoxMultiple as ScheinexamManagementIcon,
   ViewDashboard as DashboardIcon,
 } from 'mdi-material-ui';
 import React from 'react';
@@ -21,6 +21,8 @@ import AttendanceAdminView from '../view/attendance/AttendanceAdminView';
 import AttendanceView from '../view/attendance/AttendanceView';
 import CriteriaInfoView from '../view/criteria-info-view/CriteriaInfoView';
 import Dashboard from '../view/dashboard/Dashboard';
+import GenerateTutorials from '../view/generate-tutorials/GenerateTutorials';
+import ImportUsers from '../view/import-data/ImportUsers';
 import Login from '../view/Login';
 import EnterScheinexamPoints from '../view/points-scheinexam/enter-form/EnterScheinexamPoints';
 import ScheinexamPointsOverview from '../view/points-scheinexam/overview/ScheinexamPointsOverview';
@@ -38,8 +40,6 @@ import Teamoverview from '../view/teamoverview/Teamoverview';
 import TutorialManagement from '../view/tutorialmanagement/TutorialManagement';
 import TutorialSubstituteManagement from '../view/tutorialmanagement/TutorialSubstituteManagement';
 import UserManagement from '../view/usermanagement/UserManagement';
-import ImportData from '../view/import-data/ImportData';
-import GenerateTutorials from '../view/generate-tutorials/GenerateTutorials';
 
 export enum RoutingPath {
   ROOT = '/',
@@ -222,7 +222,7 @@ export const ROUTES: readonly RouteType[] = [
   {
     path: RoutingPath.IMPORT_USERS,
     title: 'Importiere Nutzer',
-    component: ImportData,
+    component: ImportUsers,
     icon: UserIcon,
     roles: [Role.ADMIN],
     isInDrawer: false,
