+++ client/src/view/import-data/adjust-data-form/AdjustImportedUserDataForm.tsx
@@ -5,7 +5,7 @@ import {
   NextStepInformation,
   useStepper,
 } from '../../../components/stepper-with-buttons/context/StepperContext';
-import { useImportDataContext } from '../ImportData.context';
+import { useImportDataContext } from '../ImportUsers.context';
 import { Formik } from 'formik';
 import FormikSelect from '../../../components/forms/components/FormikSelect';
 import FormikDebugDisplay from '../../../components/forms/components/FormikDebugDisplay';
@@ -51,7 +51,7 @@ function NO_OP() {
   /* No-Op */
 }
 
-function ImportUsers(): JSX.Element {
+function AdjustImportedUserDataForm(): JSX.Element {
   const classes = useStyles();
 
   const { setNextCallback, removeNextCallback } = useStepper();
@@ -173,4 +173,4 @@ function ImportUsers(): JSX.Element {
   );
 }
 
-export default ImportUsers;
+export default AdjustImportedUserDataForm;
