+++ client/src/view/import-data/import-csv/ImportUserCSV.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function ImportUserCSV(): JSX.Element {
+  return <div>IMPORT_CSV</div>;
+}
+
+export default ImportUserCSV;
